router-Product,Supplier,Category,Stock,Dashboard

+++ warehouse-management-system/models/dashboard.go
@@ -1,8 +1,8 @@
 package models
 
 type DashboardResponse struct {
-	TotalProducts   int `json:"total_products"`
-	TotalCategories int `json:"total_categories"`
-	TotalSuppliers  int `json:"total_suppliers"`
-	LowStockCount   int `json:"low_stock_count"`
+	TotalProducts   int64 `json:"total_products"`
+	TotalCategories int64 `json:"total_categories"`
+	TotalSuppliers  int64 `json:"total_suppliers"`
+	LowStockCount   int64 `json:"low_stock_count"`
 }
